backend/Agents/requirements.txt
langchain==0.2.16
langgraph
#langchain-openai
langchain-community
tavily-python==0.3.3
python-dotenv==1.0.1
langchain-google-genai
langchain-groq
langchain-tavily

# for testing:
langgraph-cli #
langgraph-api
langchain-core
#pip install -U "langgraph-cli[inmem]"

